Phase_II_Web_App/src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-multipart==0.0.6
bcrypt==4.1.2
pyjwt==2.8.0